--- lib/paginated_data_table_2.dart
@@ -1,585 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-// Copyright 2021 Maxim Saplin, Kristián Balaj - chnages and modifications to original Flutter implementation of PaginatedDataTable
-
-import 'dart:math' as math;
-
-import 'package:data_table_2/async_data_table_source.dart';
-import 'package:data_table_2/data_state_enum.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/gestures.dart' show DragStartBehavior;
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/widgets.dart';
-
-import 'data_table_2.dart';
-
-/// In-place replacement of standard [PaginatedDataTable] widget, mimics it API.
-/// Has the header row and paginatior always fixed to top and bottom (correspondingly).
-/// Core of the table (with data rows) is scrollable and stretching to max width/height of it's container.
-/// You can set minimal width of the table via [minWidth] property and Flex behavior of
-/// table core via [fit] property.
-/// By using [DataColumn2] instead of [DataColumn] it is possible to control
-/// relative column sizes (setting them to S, M and L). [DataRow2] provides
-/// row-level tap event handlers.
-/// See also:
-///
-///  * [DataTable2], which is not paginated.
-class PaginatedDataTable2 extends StatefulWidget {
-  /// Check out [PaginatedDataTable] for the API decription.
-  /// Key differences are [minWidth] and [fit] properties.
-  PaginatedDataTable2({
-    Key? key,
-    this.header,
-    this.actions,
-    required this.columns,
-    this.sortColumnIndex,
-    this.sortAscending = true,
-    this.onSelectAll,
-    this.dataRowHeight = kMinInteractiveDimension,
-    this.headingRowHeight = 56.0,
-    this.horizontalMargin = 24.0,
-    this.columnSpacing = 56.0,
-    this.showCheckboxColumn = true,
-    this.showFirstLastButtons = false,
-    this.initialFirstRowIndex = 0,
-    this.onPageChanged,
-    this.rowsPerPage = defaultRowsPerPage,
-    this.availableRowsPerPage = const <int>[
-      defaultRowsPerPage,
-      defaultRowsPerPage * 2,
-      defaultRowsPerPage * 5,
-      defaultRowsPerPage * 10
-    ],
-    this.onRowsPerPageChanged,
-    this.dragStartBehavior = DragStartBehavior.start,
-    required this.source,
-    this.checkboxHorizontalMargin,
-    this.wrapInCard = true,
-    this.minWidth,
-    this.fit = FlexFit.loose,
-    this.scrollController,
-    this.border,
-    this.smRatio = 0.67,
-    this.lmRatio = 1.2,
-    this.empty,
-    this.loadingWidget,
-    this.errorBuilder,
-  })  : assert(actions == null || (header != null)),
-        assert(columns.isNotEmpty),
-        assert(sortColumnIndex == null ||
-            (sortColumnIndex >= 0 && sortColumnIndex < columns.length)),
-        assert(rowsPerPage > 0),
-        assert(() {
-          if (onRowsPerPageChanged != null)
-            assert(availableRowsPerPage.contains(rowsPerPage));
-          return true;
-        }()),
-        super(key: key);
-
-  /// Displayed in case an error occurs in the [AsyncDataTableSource].
-  /// The fallback is an empty [SizedBox].
-  final Widget Function(BuildContext context, Object? error)? errorBuilder;
-
-  /// Displayed in case the [AsyncDataTableSource.getRows] is loading data.
-  /// The fallback is an empty [SizedBox].
-  final Widget? loadingWidget;
-
-  final bool wrapInCard;
-
-  /// The table card's optional header.
-  ///
-  /// This is typically a [Text] widget, but can also be a [Row] of
-  /// [TextButton]s. To show icon buttons at the top end side of the table with
-  /// a header, set the [actions] property.
-  ///
-  /// If items in the table are selectable, then, when the selection is not
-  /// empty, the header is replaced by a count of the selected items. The
-  /// [actions] are still visible when items are selected.
-  final Widget? header;
-
-  /// Icon buttons to show at the top end side of the table. The [header] must
-  /// not be null to show the actions.
-  ///
-  /// Typically, the exact actions included in this list will vary based on
-  /// whether any rows are selected or not.
-  ///
-  /// These should be size 24.0 with default padding (8.0).
-  final List<Widget>? actions;
-
-  /// The configuration and labels for the columns in the table.
-  final List<DataColumn> columns;
-
-  /// The current primary sort key's column.
-  ///
-  /// See [DataTable.sortColumnIndex].
-  final int? sortColumnIndex;
-
-  /// Whether the column mentioned in [sortColumnIndex], if any, is sorted
-  /// in ascending order.
-  ///
-  /// See [DataTable.sortAscending].
-  final bool sortAscending;
-
-  /// Invoked when the user selects or unselects every row, using the
-  /// checkbox in the heading row.
-  ///
-  /// See [DataTable.onSelectAll].
-  final ValueSetter<bool?>? onSelectAll;
-
-  /// The height of each row (excluding the row that contains column headings).
-  ///
-  /// This value is optional and defaults to kMinInteractiveDimension if not
-  /// specified.
-  final double dataRowHeight;
-
-  /// The height of the heading row.
-  ///
-  /// This value is optional and defaults to 56.0 if not specified.
-  final double headingRowHeight;
-
-  /// The horizontal margin between the edges of the table and the content
-  /// in the first and last cells of each row.
-  ///
-  /// When a checkbox is displayed, it is also the margin between the checkbox
-  /// the content in the first data column.
-  ///
-  /// This value defaults to 24.0 to adhere to the Material Design specifications.
-  ///
-  /// If [checkboxHorizontalMargin] is null, then [horizontalMargin] is also the
-  /// margin between the edge of the table and the checkbox, as well as the
-  /// margin between the checkbox and the content in the first data column.
-  final double horizontalMargin;
-
-  /// The horizontal margin between the contents of each data column.
-  ///
-  /// This value defaults to 56.0 to adhere to the Material Design specifications.
-  final double columnSpacing;
-
-  /// {@macro flutter.material.dataTable.showCheckboxColumn}
-  final bool showCheckboxColumn;
-
-  /// Flag to display the pagination buttons to go to the first and last pages.
-  final bool showFirstLastButtons;
-
-  /// The index of the first row to display when the widget is first created.
-  final int? initialFirstRowIndex;
-
-  /// Invoked when the user switches to another page.
-  ///
-  /// The value is the index of the first row on the currently displayed page.
-  final ValueChanged<int>? onPageChanged;
-
-  /// The number of rows to show on each page.
-  ///
-  /// See also:
-  ///
-  ///  * [onRowsPerPageChanged]
-  ///  * [defaultRowsPerPage]
-  final int rowsPerPage;
-
-  /// The default value for [rowsPerPage].
-  ///
-  /// Useful when initializing the field that will hold the current
-  /// [rowsPerPage], when implemented [onRowsPerPageChanged].
-  static const int defaultRowsPerPage = 10;
-
-  /// The options to offer for the rowsPerPage.
-  ///
-  /// The current [rowsPerPage] must be a value in this list.
-  ///
-  /// The values in this list should be sorted in ascending order.
-  final List<int> availableRowsPerPage;
-
-  /// Invoked when the user selects a different number of rows per page.
-  ///
-  /// If this is null, then the value given by [rowsPerPage] will be used
-  /// and no affordance will be provided to change the value.
-  final ValueChanged<int?>? onRowsPerPageChanged;
-
-  /// The data source which provides data to show in each row.
-  ///
-  /// This object should generally have a lifetime longer than the
-  /// [PaginatedDataTable2] widget itself; it should be reused each time the
-  /// [PaginatedDataTable2] constructor is called.
-  final AsyncDataTableSource source;
-
-  /// {@macro flutter.widgets.scrollable.dragStartBehavior}
-  final DragStartBehavior dragStartBehavior;
-
-  /// Horizontal margin around the checkbox, if it is displayed.
-  ///
-  /// If null, then [horizontalMargin] is used as the margin between the edge
-  /// of the table and the checkbox, as well as the margin between the checkbox
-  /// and the content in the first data column. This value defaults to 24.0.
-  final double? checkboxHorizontalMargin;
-
-  /// If set, the table will stop shrinking below the threshold and provide
-  /// horizontal scrolling. Useful for the cases with narrow screens (e.g. portrait phone orientation)
-  /// and lots of columns (that get messed with little space)
-  // TODO add test
-  final double? minWidth;
-
-  /// Data rows are wrapped in Flexible widget, this property sets its' fit property.
-  /// When ther're few rows it determines if the core
-  /// of the table must grow and fill the contrainer (FlexFit.tight - useful if
-  /// you want the paginator to stick to the bottom when there're few rows) or
-  /// of you want to have the table to take minimal space and do not have bottom
-  /// pager stick to the bottom (FlexFit.loose)
-  // TODO add test
-  final FlexFit fit;
-
-  // TODO: Add test
-  /// Set vertical and horizontal borders between cells, as well as outside borders around table.
-  /// NOTE: setting this field will disable standard horizontal dividers which are controlled by
-  /// themes and [dividerThickness] property
-  final TableBorder? border;
-
-  /// Placeholder widget which is displayed whenever the data rows are empty.
-  /// The widget will be displayed below column
-  /// Fallback is the [SizedBox].
-  final Widget? empty;
-
-  // TODO: Add test
-  /// Determines ratio of Small column's width to Medium column's width.
-  /// I.e. 0.5 means that Small column is twice narower than Medium column.
-  final double smRatio;
-
-  // TODO: Add test
-  /// Determines ratio of Large column's width to Medium column's width.
-  /// I.e. 2.0 means that Large column is twice wider than Medium column.
-  final double lmRatio;
-
-  /// Exposes scroll controller of the SingleChildScrollView that makes data rows horizontally scrollable
-  // TODO add test
-  final ScrollController? scrollController;
-
-  @override
-  PaginatedDataTable2State createState() => PaginatedDataTable2State();
-}
-
-/// Holds the state of a [PaginatedDataTable2].
-///
-/// The table can be programmatically paged using the [pageTo] method.
-class PaginatedDataTable2State extends State<PaginatedDataTable2> {
-  late int _firstRowIndex;
-  late int _rowCount;
-  late bool _rowCountApproximate;
-  int _selectedRowCount = 0;
-  final Map<int, DataRow?> _rows = <int, DataRow?>{};
-
-  @override
-  void initState() {
-    super.initState();
-    _firstRowIndex = PageStorage.of(context)?.readState(context) as int? ??
-        widget.initialFirstRowIndex ??
-        0;
-    widget.source.addListener(_handleDataSourceChanged);
-    _handleDataSourceChanged();
-  }
-
-  @override
-  void didUpdateWidget(PaginatedDataTable2 oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.source != widget.source) {
-      oldWidget.source.removeListener(_handleDataSourceChanged);
-      widget.source.addListener(_handleDataSourceChanged);
-      _handleDataSourceChanged();
-    }
-  }
-
-  @override
-  void dispose() {
-    widget.source.removeListener(_handleDataSourceChanged);
-    super.dispose();
-  }
-
-  void _handleDataSourceChanged() {
-    setState(() {
-      _rowCount = widget.source.rowCount;
-      _rowCountApproximate = widget.source.isRowCountApproximate;
-      _selectedRowCount = widget.source.selectedRowCount;
-      _rows.clear();
-    });
-  }
-
-  /// Ensures that the given row is visible.
-  void pageTo(int rowIndex) {
-    final int oldFirstRowIndex = _firstRowIndex;
-    setState(() {
-      final int rowsPerPage = widget.rowsPerPage;
-      _firstRowIndex = (rowIndex ~/ rowsPerPage) * rowsPerPage;
-    });
-    if ((widget.onPageChanged != null) && (oldFirstRowIndex != _firstRowIndex))
-      widget.onPageChanged!(_firstRowIndex);
-  }
-
-  Future<List<DataRow>> _getRows(int firstRowIndex, int rowsPerPage) {
-    final List<DataRow> result = <DataRow>[];
-
-    if (widget.empty != null && widget.source.rowCount < 1)
-      return SynchronousFuture(result);
-
-    final int nextPageFirstRowIndex = firstRowIndex + rowsPerPage;
-
-    return widget.source.getRows(firstRowIndex, nextPageFirstRowIndex - 1);
-  }
-
-  void _handleFirst() {
-    pageTo(0);
-  }
-
-  void _handlePrevious() {
-    pageTo(math.max(_firstRowIndex - widget.rowsPerPage, 0));
-  }
-
-  void _handleNext() {
-    pageTo(_firstRowIndex + widget.rowsPerPage);
-  }
-
-  void _handleLast() {
-    pageTo(((_rowCount - 1) / widget.rowsPerPage).floor() * widget.rowsPerPage);
-  }
-
-  bool _isNextPageUnavailable() =>
-      !_rowCountApproximate &&
-      (_firstRowIndex + widget.rowsPerPage >= _rowCount);
-
-  final GlobalKey _tableKey = GlobalKey();
-
-  @override
-  Widget build(BuildContext context) {
-    // TODO(ianh): This whole build function doesn't handle RTL yet.
-    assert(debugCheckHasMaterialLocalizations(context));
-    final ThemeData themeData = Theme.of(context);
-    final MaterialLocalizations localizations =
-        MaterialLocalizations.of(context);
-    // HEADER
-    final List<Widget> headerWidgets = <Widget>[];
-    double startPadding = widget.horizontalMargin;
-    if (_selectedRowCount == 0 && widget.header != null) {
-      headerWidgets.add(Expanded(child: widget.header!));
-      if (widget.header is ButtonBar) {
-        // We adjust the padding when a button bar is present, because the
-        // ButtonBar introduces 2 pixels of outside padding, plus 2 pixels
-        // around each button on each side, and the button itself will have 8
-        // pixels internally on each side, yet we want the left edge of the
-        // inside of the button to line up with the 24.0 left inset.
-        startPadding = 12.0;
-      }
-    } else if (widget.header != null) {
-      headerWidgets.add(Expanded(
-        child: Text(localizations.selectedRowCountTitle(_selectedRowCount)),
-      ));
-    }
-    if (widget.actions != null) {
-      headerWidgets.addAll(widget.actions!.map<Widget>((Widget action) {
-        return Padding(
-          // 8.0 is the default padding of an icon button
-          padding: const EdgeInsetsDirectional.only(start: 24.0 - 8.0 * 2.0),
-          child: action,
-        );
-      }).toList());
-    }
-
-    // FOOTER
-    final TextStyle? footerTextStyle = themeData.textTheme.caption;
-    final List<Widget> footerWidgets = <Widget>[];
-    if (widget.onRowsPerPageChanged != null) {
-      final List<Widget> availableRowsPerPage = widget.availableRowsPerPage
-          .where(
-              (int value) => value <= _rowCount || value == widget.rowsPerPage)
-          .map<DropdownMenuItem<int>>((int value) {
-        return DropdownMenuItem<int>(
-          value: value,
-          child: Text('$value'),
-        );
-      }).toList();
-      footerWidgets.addAll(<Widget>[
-        Container(
-            width:
-                14.0), // to match trailing padding in case we overflow and end up scrolling
-        Text(localizations.rowsPerPageTitle),
-        ConstrainedBox(
-          constraints: const BoxConstraints(
-              minWidth: 64.0), // 40.0 for the text, 24.0 for the icon
-          child: Align(
-            alignment: AlignmentDirectional.centerEnd,
-            child: DropdownButtonHideUnderline(
-              child: DropdownButton<int>(
-                items: availableRowsPerPage.cast<DropdownMenuItem<int>>(),
-                value: widget.rowsPerPage,
-                onChanged: widget.onRowsPerPageChanged,
-                style: footerTextStyle,
-                iconSize: 24.0,
-              ),
-            ),
-          ),
-        ),
-      ]);
-    }
-    footerWidgets.addAll(<Widget>[
-      Container(width: 32.0),
-      Text(
-        localizations.pageRowsInfoTitle(
-          _firstRowIndex + 1,
-          _firstRowIndex + widget.rowsPerPage,
-          _rowCount,
-          _rowCountApproximate,
-        ),
-      ),
-      Container(width: 32.0),
-      if (widget.showFirstLastButtons)
-        IconButton(
-          icon: const Icon(Icons.skip_previous),
-          padding: EdgeInsets.zero,
-          //tooltip: localizations.firstPageTooltip,
-          onPressed: _firstRowIndex <= 0 ? null : _handleFirst,
-        ),
-      IconButton(
-        icon: const Icon(Icons.chevron_left),
-        padding: EdgeInsets.zero,
-        tooltip: localizations.previousPageTooltip,
-        onPressed: _firstRowIndex <= 0 ? null : _handlePrevious,
-      ),
-      Container(width: 24.0),
-      IconButton(
-        icon: const Icon(Icons.chevron_right),
-        padding: EdgeInsets.zero,
-        tooltip: localizations.nextPageTooltip,
-        onPressed: _isNextPageUnavailable() ? null : _handleNext,
-      ),
-      if (widget.showFirstLastButtons)
-        IconButton(
-          icon: const Icon(Icons.skip_next),
-          padding: EdgeInsets.zero,
-          //tooltip: localizations.lastPageTooltip,
-          onPressed: _isNextPageUnavailable() ? null : _handleLast,
-        ),
-      Container(width: 14.0),
-    ]);
-
-    // CARD
-    return LayoutBuilder(
-      builder: (BuildContext context, BoxConstraints constraints) {
-        Widget t = Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: <Widget>[
-            if (headerWidgets.isNotEmpty)
-              Semantics(
-                container: true,
-                child: DefaultTextStyle(
-                  // These typographic styles aren't quite the regular ones. We pick the closest ones from the regular
-                  // list and then tweak them appropriately.
-                  // See https://material.io/design/components/data-tables.html#tables-within-cards
-                  style: _selectedRowCount > 0
-                      ? themeData.textTheme.subtitle1!
-                          .copyWith(color: themeData.accentColor)
-                      : themeData.textTheme.headline6!
-                          .copyWith(fontWeight: FontWeight.w400),
-                  child: IconTheme.merge(
-                    data: const IconThemeData(opacity: 0.54),
-                    child: Ink(
-                      height: 64.0,
-                      color: _selectedRowCount > 0
-                          ? themeData.secondaryHeaderColor
-                          : null,
-                      child: Padding(
-                        padding: EdgeInsetsDirectional.only(
-                            start: startPadding, end: 14.0),
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.end,
-                          children: headerWidgets,
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            Flexible(
-              fit: widget.fit,
-              child: ConstrainedBox(
-                constraints: BoxConstraints(minWidth: constraints.minWidth),
-                child: FutureBuilder<List<DataRow>>(
-                  future: _getRows(_firstRowIndex, widget.rowsPerPage),
-                  builder: (
-                    BuildContext context,
-                    AsyncSnapshot<List<DataRow>> snapshot,
-                  ) {
-                    DataState state = () {
-                      switch (snapshot.connectionState) {
-                        case ConnectionState.waiting:
-                          return DataState.loading;
-                        default:
-                          if (snapshot.hasError)
-                            return DataState.error;
-                          else
-                            return DataState.done;
-                      }
-                    }();
-
-                    return DataTable2(
-                      key: _tableKey,
-                      columns: widget.columns,
-                      sortColumnIndex: widget.sortColumnIndex,
-                      sortAscending: widget.sortAscending,
-                      onSelectAll: widget.onSelectAll,
-                      // Make sure no decoration is set on the DataTable
-                      // from the theme, as its already wrapped in a Card.
-                      decoration: const BoxDecoration(),
-                      dataRowHeight: widget.dataRowHeight,
-                      headingRowHeight: widget.headingRowHeight,
-                      horizontalMargin: widget.horizontalMargin,
-                      //TODO - fix when Flutter 2.1.0 goes stable
-                      //checkboxHorizontalMargin: widget.checkboxHorizontalMargin,
-                      columnSpacing: widget.columnSpacing,
-                      showCheckboxColumn: widget.showCheckboxColumn,
-                      showBottomBorder: true,
-                      minWidth: widget.minWidth,
-                      scrollController: widget.scrollController,
-                      empty: widget.empty,
-                      border: widget.border,
-                      smRatio: widget.smRatio,
-                      lmRatio: widget.lmRatio,
-                      rows: snapshot.data ?? [],
-                      loadingWidget: widget.loadingWidget,
-                      errorBuilder: (context) =>
-                          widget.errorBuilder?.call(context, snapshot.error) ??
-                          SizedBox(),
-                      dataState: state,
-                    );
-                  },
-                ),
-              ),
-            ),
-            DefaultTextStyle(
-              style: footerTextStyle!,
-              child: IconTheme.merge(
-                data: const IconThemeData(opacity: 0.54),
-                child: SizedBox(
-                  // TODO(bkonyi): this won't handle text zoom correctly,
-                  //  https://github.com/flutter/flutter/issues/48522
-                  height: 56.0,
-                  child: SingleChildScrollView(
-                    dragStartBehavior: widget.dragStartBehavior,
-                    scrollDirection: Axis.horizontal,
-                    reverse: true,
-                    child: Row(
-                      children: footerWidgets,
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ],
-        );
-
-        if (widget.wrapInCard) t = Card(semanticContainer: false, child: t);
-
-        return t;
-      },
-    );
-  }
-}
