+++ lib/paginated_data_tables/paginated_data_table_2.dart
@@ -4,16 +4,25 @@
 
 // Copyright 2021 Maxim Saplin, Kristián Balaj - changes and modifications to original Flutter implementation of PaginatedDataTable
 
-import 'package:data_table_2/async_data_table_source.dart';
-import 'package:data_table_2/data_state_enum.dart';
-import 'package:data_table_2/paged_data_table_2.dart';
-import 'package:data_table_2/paged_data_table_2_base.dart';
-import 'package:flutter/foundation.dart';
+import 'package:data_table_2/paginated_data_tables/paginated_data_table_2_base.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
-class PagedDataTable2Async extends PagedDataTable2Base<AsyncDataTableSource> {
-  PagedDataTable2Async({
+/// In-place replacement of standard [PaginatedDataTable] widget, mimics it API.
+/// Has the header row and paginatior always fixed to top and bottom (correspondingly).
+/// Core of the table (with data rows) is scrollable and stretching to max width/height of it's container.
+/// You can set minimal width of the table via [minWidth] property and Flex behavior of
+/// table core via [fit] property.
+/// By using [DataColumn2] instead of [DataColumn] it is possible to control
+/// relative column sizes (setting them to S, M and L). [DataRow2] provides
+/// row-level tap event handlers.
+/// See also:
+///
+///  * [DataTable2], which is not paginated.
+class PaginatedDataTable2 extends PaginatedDataTable2Base<DataTableSource> {
+  /// Check out [PaginatedDataTable] for the API decription.
+  /// Key differences are [minWidth] and [fit] properties.
+  PaginatedDataTable2({
     Key? key,
     Widget? header,
     List<Widget>? actions,
@@ -29,16 +38,16 @@ class PagedDataTable2Async extends PagedDataTable2Base<AsyncDataTableSource> {
     bool showFirstLastButtons = false,
     int? initialFirstRowIndex = 0,
     ValueChanged<int>? onPageChanged,
-    int rowsPerPage = PagedDataTable2Base.defaultRowsPerPage,
+    int rowsPerPage = PaginatedDataTable2Base.defaultRowsPerPage,
     List<int> availableRowsPerPage = const <int>[
-      PagedDataTable2Base.defaultRowsPerPage,
-      PagedDataTable2Base.defaultRowsPerPage * 2,
-      PagedDataTable2Base.defaultRowsPerPage * 5,
-      PagedDataTable2Base.defaultRowsPerPage * 10
+      PaginatedDataTable2Base.defaultRowsPerPage,
+      PaginatedDataTable2Base.defaultRowsPerPage * 2,
+      PaginatedDataTable2Base.defaultRowsPerPage * 5,
+      PaginatedDataTable2Base.defaultRowsPerPage * 10
     ],
     ValueChanged<int?>? onRowsPerPageChanged,
     DragStartBehavior dragStartBehavior = DragStartBehavior.start,
-    required AsyncDataTableSource dataSource,
+    required DataTableSource dataSource,
     double? checkboxHorizontalMargin,
     bool wrapInCard = true,
     double? minWidth,
@@ -48,8 +57,6 @@ class PagedDataTable2Async extends PagedDataTable2Base<AsyncDataTableSource> {
     double smRatio = 0.67,
     double lmRatio = 1.2,
     Widget? empty,
-    this.loadingWidget,
-    this.errorBuilder,
   }) : super(
           header: header,
           actions: actions,
@@ -80,20 +87,12 @@ class PagedDataTable2Async extends PagedDataTable2Base<AsyncDataTableSource> {
           empty: empty,
         );
 
-  /// Displayed in case an error occurs in the [AsyncDataTableSource].
-  /// The fallback is an empty [SizedBox].
-  final Widget Function(BuildContext context, Object? error)? errorBuilder;
-
-  /// Widget that is displayed in case the data rows are being loaded.
-  /// The fallback is an empty [SizedBox].
-  final Widget? loadingWidget;
-
   @override
-  PagedDataTable2BaseState createState() => PagedDataTable2State();
+  PaginatedDataTable2BaseState createState() => PaginatedDataTable2State();
 }
 
-class PagedDataTable2AsyncState
-    extends PagedDataTable2BaseState<PagedDataTable2Async> {
+class PaginatedDataTable2State
+    extends PaginatedDataTable2BaseState<PaginatedDataTable2> {
   @override
   bool get dataSourceIsRowCountApproximate =>
       widget.dataSource.isRowCountApproximate;
@@ -104,46 +103,8 @@ class PagedDataTable2AsyncState
   @override
   int get dataSourceSelectedRowCount => widget.dataSource.selectedRowCount;
 
-  Future<List<DataRow>> _getRows(int firstRowIndex, int rowsPerPage) {
-    final List<DataRow> result = <DataRow>[];
-
-    if (widget.empty != null && widget.dataSource.rowCount < 1)
-      return SynchronousFuture(result);
-
-    final int nextPageFirstRowIndex = firstRowIndex + rowsPerPage;
-
-    return widget.dataSource.getRows(firstRowIndex, nextPageFirstRowIndex - 1);
-  }
-
   @override
   Widget createDataTableContextWidget() {
-    return FutureBuilder<List<DataRow>>(
-      future: _getRows(firstRowIndex, widget.rowsPerPage),
-      builder: (
-        BuildContext context,
-        AsyncSnapshot<List<DataRow>> snapshot,
-      ) {
-        DataState state = () {
-          switch (snapshot.connectionState) {
-            case ConnectionState.waiting:
-              return DataState.loading;
-            default:
-              if (snapshot.hasError)
-                return DataState.error;
-              else
-                return DataState.done;
-          }
-        }();
-
-        return createDataTableWidget(
-          rows: snapshot.data ?? [],
-          errorBuilder: (context) =>
-              widget.errorBuilder?.call(context, snapshot.error) ??
-              const SizedBox(),
-          state: state,
-          loadingWidget: widget.loadingWidget ?? const SizedBox(),
-        );
-      },
-    );
+    return Container();
   }
 }
